[IMP] website, *: allow selecting a single record in dynamic snippets
*: website_blog, website_event, website_sale

Currently, dynamic snippets only support filters to display records.
This commit improves the current implementation, and allows users to
select a single record directly in the snippet (which should
automatically switch to the "single record" mode, when the number of
records to fetch is set to "one").

- Generic snippets can switch between models, while specific ones
will limit selection to their related model.

- A loading effect was added while fetching snippet content.

Remarks:

- The single record mode was enabled on "blog" and "event" snippets.
The "appointment"/"product" templates will be added as part of a
separate refactoring.

- This commit also removes the 'Template' option from dynamic snippets.
Instead, the various snippet layouts are now directly accessible from
the snippet selection dialog (For product snippets, this feature will
be enabled right after the snippets design refactoring).

task-4280375

Part-of: odoo/odoo#190514
Related: odoo/enterprise#75556
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>
Co-authored-by: Antoine (anso) <anso@odoo.com>